Add functionality to the upload button
if the button is being clicked while on the homepage, the dropzone will be auto clicked programmatically and the file browser will pop up to the user allowing them to add files. If the button is not clicked on the homepage, it redirects them to the homepage where you can upload images there.