Implement client-side encryption on file uploads
This commit utilizes NodeJS's built in crypto module to be able to encrypt and decrypt file buffers on the client side with a password that is only known to the user that uploads the content. No file-readable data, passwords or encryption processes take place on any servers or third parties. It is all done within the clients browser, which prevents any third parties such as any server administrators, Vercel or AWS administrators from viewing the file. It is then decrypted only on the client side when viewing the file, which then prompts the user for the password used to encrypt the file. The file is then displayed on the client side if the correct password is given for decryption, if it fails to load, that means the client entered the wrong password and it then re-prompts the user to enter the correct one.