Implement an API key system
This commit allows clients to generate and revoke API keys. The purpose of this is so that they can configure an external image upload client for example ShareX to use their accounts API key, which identifies that the uploads made from them are tied to a certain account. This allows the user to be able to go back into their gallery on their account and view the past uploads they made from the external client.