Update image viewer to reference by a unique value such as a bucket key
Initially it was referencing by the actual object and not the unique bucket key, which was due to confusing variable names. This has been corrected so that it references the actual bucket key value and variable names have been re-named appropriately to avoid an issue like this in the future. Additionally, the functionality of how it processes retrieving encrypted buffers has been refactored to reduce overhead, so that it only makes one call to fetch the encrypted buffer, and not a call every-time you get your password incorrect. The decryption function as been isolated into its own one instead of being bundled with the download function