Accommodate encrypted strings to be 3000 characters long
Originally, it would do the correct checks in preventing plaintext notes from being 3000 characters long, but since an encrypted version of the string is larger in length, it has to consider that large length. So I tested how many characters an encrypted string with 3000 characters as plaintext would be and the result is always 6032 regardless of the characters. This commit makes the serverless API route check if the content is below 6032 characters and if not it throws an error telling the user that the encrypted string has to be below 3000 characters